Refactors RaflesList to component
