+++ pages/index.js
@@ -3,61 +3,19 @@ import { Card, Button } from 'semantic-ui-react';
 import factory from '../ethereum/factory';
 import Raffle from '../ethereum/raffle';
 import Layout from '../components/Layout';
+import RafflesList from '../components/RafflesList';
 import { Link } from '../routes';
 class RaffleIndex extends Component {
-	state = { raffles: [] };
-	async componentWillMount() {
-		const getRaffles = await factory.methods.getDeployedRaffles().call();
-		const raffles = getRaffles.map(async address => {
-			const raffle = Raffle(address);
-			const res = await raffle.methods.getRaffleSummary().call();
-			// console.log(res);
-			const raffleObj = await {
-				raffleAddress: res[0],
-				raffleBalance: res[1],
-				soldTickets: res[2],
-				remainingNumberOfTickets: res[3],
-				ticketsBlock: res[4],
-				soldTicketsNumbers: res[5]
-			};
-			await this.state.raffles.push(raffleObj);
-		});
-	}
-
-	// static async getInitialProps() {
-	// 	return { raffles: rafflesArr };
-	// }
-
-	renderRaffles() {
-		console.log('this.state ', this.state);
-		const items = this.state.raffles.map((raffle, index) => {
-			return {
-				header: raffle.raffleAddress,
-				description: (
-					<div>
-						<ul>
-							<li>Balance: {raffle.raffleBalance} ETH</li>
-							<li>
-								Remaining Number of Tickets: {raffle.remainingNumberOfTickets}
-							</li>
-						</ul>
-						<Link route={`/raffles/${raffle.raffleAddress}`}>
-							<a>View raffle</a>
-						</Link>
-					</div>
-				),
-				fluid: true
-			};
-		});
-
-		return <Card.Group keys={items.id} items={items} />;
-	}
-
 	render() {
 		return (
 			<Layout>
 				<div>
 					<h3>Open raffles</h3>
+					<Link route="/">
+						<a>
+							<Button content="See all raffles" icon="eye" primary={true} />
+						</a>
+					</Link>
 					<Link route="/raffles/new">
 						<a>
 							<Button
@@ -69,7 +27,7 @@ class RaffleIndex extends Component {
 						</a>
 					</Link>
 
-					{this.renderRaffles()}
+					<RafflesList />
 				</div>
 			</Layout>
 		);
